refactor(settings): 二级页拆分搬家，删 Settings 容器与 4-tab 结构
