+++ src/sidepanel/components/Settings.localbridge.test.tsx
@@ -8,7 +8,7 @@
  */
 import { describe, it, expect, afterEach } from "vitest";
 import { render, screen, fireEvent, cleanup, waitFor } from "@testing-library/react";
-import { LocalBridgeSection } from "./Settings";
+import { LocalBridgeSection } from "./settings/pages/BridgePage";
 import { chromeMock } from "@/test/setup";
 
 const GRANT = {
