--- src/sidepanel/components/Settings.tsx
@@ -1,831 +1,0 @@
-import { useState, useEffect, useCallback, useRef } from "react";
-import type { ProviderRef, BuiltinProvider, ModelMeta } from "@/lib/model-router";
-import {
-  createInstance, listInstances, deleteInstance,
-  updateInstance, firstModelForProvider,
-  type DecryptedInstance,
-} from "@/lib/instances";
-import {
-  getProviderCustomModels,
-  addProviderCustomModel,
-  removeProviderCustomModel,
-} from "@/lib/provider-custom-models";
-import {
-  getProviderCustomModelMetas,
-  setProviderCustomModelMeta,
-  removeProviderCustomModelMeta,
-  type StoredCustomModelMeta,
-} from "@/lib/provider-custom-model-meta";
-import { getProviderMeta, resolveProviderMeta, resolveEndpointVariant } from "@/lib/model-router/providers/registry";
-import { fetchOpenRouterModels } from "@/lib/openrouter-models-fetch";
-import { isCdpInputEnabled, setCdpInputEnabled } from "@/lib/cdp-input-enabled";
-import {
-  addCustomProviderModel, updateCustomProviderModel, removeCustomProviderModel,
-  CUSTOM_PREFIX, providerRefToId, listCustomProviders,
-} from "@/lib/custom-providers";
-import { IconButton } from "./ui/IconButton";
-import SkillsList from "./SkillsList";
-import SearchProviderSection from "./SearchProviderSection";
-import InstanceForm, { type InstanceFormPayload } from "./InstanceForm";
-import InstancesList from "./InstancesList";
-import NewConfigWizard from "./NewConfigWizard";
-import type { ProviderTestOptions } from "./NewConfigWizard";
-import { Collapse } from "./ui/Collapse";
-import AssistantLanguageSelect from "./AssistantLanguageSelect";
-import LanguageSelect from "./LanguageSelect";
-import { useT, getLocale } from "@/lib/i18n";
-import { buildGithubNewIssueUrl, buildFeedbackMailto, type FeedbackEnv } from "@/lib/feedback";
-import { testProviderConnection } from "@/lib/provider-test";
-import { submitFeedback } from "@/lib/managed-account";
-import { readRecentLogs } from "@/lib/log-buffer";
-import { capLogBytes } from "@/lib/log-cap";
-import type { GrantRecord, AuditEntry } from "@/types/local-bridge";
-
-interface Props {
-  onBack: () => void;
-  onRunSkill?: (skillId: string, skillName: string) => void;
-  /** Bumped by a website "Subscribe" deep-link — opens the New Config wizard in
-   *  managed mode (the official-subscription screen). */
-  openSubscribeNonce?: number;
-}
-
-type Tab = "configs" | "skills" | "search" | "general";
-
-export default function Settings({ onBack, onRunSkill, openSubscribeNonce }: Props) {
-  const t = useT();
-  const [tab, setTab] = useState<Tab>("configs");
-  const [instances, setInstances] = useState<DecryptedInstance[]>([]);
-  const [expandedId, setExpandedId] = useState<string | null>(null);
-  const [showWizard, setShowWizard] = useState(false);
-  const [cdpInput, setCdpInput] = useState<boolean | undefined>(undefined);
-  const [testResult, setTestResult] = useState<Record<string, { ok: boolean; message: string }>>({});
-  const [testingIds, setTestingIds] = useState<Record<string, boolean>>({});
-  const testingIdsRef = useRef<Set<string>>(new Set());
-  // Per-provider custom models pool — sticky across instances of the same provider.
-  const [providerPools, setProviderPools] = useState<Record<string, string[]>>({});
-  // Per-provider custom model meta (vision, maxContextTokens) keyed by provider then modelId.
-  const [providerMetas, setProviderMetas] = useState<Record<string, Record<string, StoredCustomModelMeta>>>({});
-  const [customProviderNames, setCustomProviderNames] = useState<Record<string, string>>({});
-
-  const reload = useCallback(async () => {
-    const list = await listInstances();
-    setInstances(list);
-    // Refresh pool for every provider currently represented in the instance list.
-    const providers = Array.from(new Set(list.map((i) => i.provider)));
-    const pools = await Promise.all(providers.map((p) => getProviderCustomModels(p).then((v) => [p, v] as const)));
-    setProviderPools(Object.fromEntries(pools));
-    // pcmm metas are builtin-scoped; filter out custom: providers (the cast is then safe).
-    const builtinProviders = providers.filter((p) => !p.startsWith(CUSTOM_PREFIX));
-    const metas = await Promise.all(
-      builtinProviders.map((p) => getProviderCustomModelMetas(p as BuiltinProvider).then((v) => [p, v] as const)),
-    );
-    setProviderMetas(Object.fromEntries(metas));
-    const customProviders = await listCustomProviders();
-    setCustomProviderNames(
-      Object.fromEntries(customProviders.map((cp) => [`${CUSTOM_PREFIX}${cp.id}`, cp.name])),
-    );
-  }, []);
-
-  useEffect(() => {
-    reload();
-    isCdpInputEnabled().then(setCdpInput);
-  }, [reload]);
-
-  // Website "Subscribe" deep-link → land on the Configs tab with the New Config
-  // wizard open; NewConfigWizard reads the same nonce to switch to managed mode.
-  useEffect(() => {
-    if (!openSubscribeNonce) return;
-    setTab("configs");
-    setShowWizard(true);
-  }, [openSubscribeNonce]);
-
-  async function handleCreate(provider: ProviderRef, payload: InstanceFormPayload) {
-    await createInstance({ provider, ...payload });
-    setShowWizard(false);
-    await reload();
-  }
-
-  async function handleSaveEdit(id: string, payload: InstanceFormPayload) {
-    const patch: { apiKey?: string; endpointVariant: string | null } = {
-      // undefined = 用户选了默认端点 → null 显式清除存储字段
-      endpointVariant: payload.endpointVariant ?? null,
-    };
-    // Only re-encrypt the key if the user actually typed a new one.
-    // An empty apiKey means "keep existing" — do NOT pass it to updateInstance.
-    if (payload.apiKey.trim().length > 0) patch.apiKey = payload.apiKey;
-    await updateInstance(id, patch);
-    setExpandedId(null); // collapse after save
-    await reload();
-  }
-
-  async function handleDelete(id: string) {
-    if (!confirm(t("settings.forgetConfirm"))) return;
-    await deleteInstance(id);
-    setExpandedId(null);
-    await reload();
-  }
-
-  async function handleTest(
-    id: string | null,
-    provider: ProviderRef,
-    payload: InstanceFormPayload,
-    options: ProviderTestOptions = {},
-  ) {
-    const key = id ?? "_new";
-    if (testingIdsRef.current.has(key)) return;
-    testingIdsRef.current.add(key);
-    setTestingIds((p) => ({ ...p, [key]: true }));
-    setTestResult((p) => {
-      const next = { ...p };
-      delete next[key];
-      return next;
-    });
-
-    const meta = await resolveProviderMeta(provider) ?? draftProviderMeta(provider, options);
-    if (!meta) {
-      setTestResult((p) => ({ ...p, [key]: { ok: false, message: `Unknown provider: ${provider}` } }));
-      testingIdsRef.current.delete(key);
-      setTestingIds((p) => ({ ...p, [key]: false }));
-      return;
-    }
-    // 端点与模型池跟随表单里未保存的 variant 选择（而非存量 instance 字段）；
-    // 兜底也传 variantOverride（null=强制默认池），避免读到存量 variant 的模型与 baseUrl 不同源
-    const variant = resolveEndpointVariant(meta, payload.endpointVariant);
-    const inst = id ? instances.find((i) => i.id === id) : undefined;
-    const model = payload.customModels[0]
-      ?? variant?.models?.[0]?.id
-      ?? meta.models[0]?.id
-      ?? options.candidateModels?.[0]?.id
-      ?? inst?.fetchedModels?.[0]?.id
-      ?? (await firstModelForProvider(provider, id ?? undefined, payload.endpointVariant ?? null))
-      ?? "";
-    const cfg = {
-      provider,
-      model,
-      // If apiKey is empty (edit mode, user didn't retype), fall back to instance's stored key
-      apiKey: payload.apiKey.trim() || (() => {
-        if (!id) return payload.apiKey;
-        return inst?.apiKey ?? payload.apiKey;
-      })(),
-      baseUrl: ((options.baseUrl?.trim() || variant?.baseUrl) ?? meta.defaultBaseUrl).replace(/\/+$/, ""),
-      providerName: options.providerName ?? meta.name,
-    };
-    try {
-      if (!cfg.apiKey.trim()) throw new Error("API key cannot be empty");
-      if (!cfg.model.trim()) throw new Error("No model available for test");
-      await testProviderConnection(cfg);
-      setTestResult((p) => ({ ...p, [key]: { ok: true, message: "" } }));
-    } catch (e) {
-      setTestResult((p) => ({ ...p, [key]: { ok: false, message: e instanceof Error ? e.message : "Failed" } }));
-    } finally {
-      testingIdsRef.current.delete(key);
-      setTestingIds((p) => ({ ...p, [key]: false }));
-    }
-  }
-
-  return (
-    <div className="flex h-full flex-col">
-      <header className="flex flex-shrink-0 items-center gap-2.5 border-b border-line bg-canvas px-3.5 py-3">
-        <IconButton
-          onClick={onBack}
-          size="sm"
-          variant="ghost"
-          aria-label={t("settings.backToAgent")}
-          icon={
-            <svg width="14" height="14" viewBox="0 0 14 14" fill="none">
-              <path d="M9 11L5 7L9 3" stroke="currentColor" strokeWidth="1.25" strokeLinecap="round" strokeLinejoin="round" />
-            </svg>
-          }
-        />
-        <span className="text-[17px] font-semibold tracking-[-0.01em] text-fg-1">{t("settings.title")}</span>
-      </header>
-      <div className="flex-shrink-0 border-b border-line bg-canvas px-3.5 pb-3.5 pt-3">
-        <SegmentedTabs value={tab} onChange={setTab} />
-      </div>
-
-      <div className="flex-1 overflow-y-auto px-4 py-6">
-        <div key={tab} className="view-enter">
-        {tab === "configs" ? (
-          <div className="flex flex-col gap-7">
-            <section className="flex flex-col gap-3.5">
-              <div className="flex items-center justify-between gap-3">
-                <span className="text-[16px] font-semibold tracking-[-0.01em] text-fg-1">{t("settings.myConfigs.title")}</span>
-                {!showWizard && (
-                  <button
-                    onClick={() => setShowWizard(true)}
-                    className="flex h-8 items-center gap-2 rounded-control border border-line bg-transparent px-3 text-[12px] text-accent transition-colors hover:bg-field"
-                  >
-                    {t("settings.myConfigs.newConfigButton")}
-                  </button>
-                )}
-              </div>
-
-              {/* 新增配置表单：置于列表上方，展开时 height+淡入 把已配置列表推下去 */}
-              <Collapse open={showWizard}>
-                <NewConfigWizard
-                  onCreate={handleCreate}
-                  onTest={(p, payload, options) => handleTest(null, p, payload, options)}
-                  existingProviderRefs={instances.map((i) => i.provider)}
-                  testing={!!testingIds["_new"]}
-                  testResult={testResult["_new"] ?? null}
-                  onCancel={() => setShowWizard(false)}
-                  managedEntryNonce={openSubscribeNonce}
-                />
-              </Collapse>
-
-              {instances.length === 0 && !showWizard && (
-                <div className="rounded-control border border-accent-line bg-accent-tint px-3 py-2 text-[12px] leading-5 text-fg-1">
-                  {t("settings.myConfigs.emptyBanner")}
-                </div>
-              )}
-
-              <InstancesList
-                instances={instances}
-                customProviderNames={customProviderNames}
-                expandedId={expandedId}
-                onToggleExpand={(id) => setExpandedId(expandedId === id ? null : id)}
-                renderForm={(id) => {
-                  const inst = instances.find((i) => i.id === id)!;
-                  const result = testResult[id];
-                  // Merge per-instance customModels (back-compat) with the
-                  // per-provider sticky pool so newly-typed ids show up across
-                  // instances of the same provider.
-                  const pool = providerPools[inst.provider] ?? [];
-                  const mergedCustomModels = Array.from(
-                    new Set([...(inst.customModels ?? []), ...pool]),
-                  );
-                  // Custom-provider models live on the provider entity; builtin
-                  // custom models live in the pcm pool + pcmm sidecar. The model
-                  // callbacks below route by provider type. `bp`/`cpId` are only
-                  // dereferenced on their matching branch, so the casts are safe.
-                  const isCustom = inst.provider.startsWith(CUSTOM_PREFIX);
-                  const bp = inst.provider as BuiltinProvider;
-                  const cpId = providerRefToId(inst.provider);
-                  return (
-                    <>
-                      <InstanceForm
-                        mode="edit"
-                        provider={inst.provider}
-                        initialNickname={inst.nickname}
-                        initialEndpointVariant={inst.endpointVariant}
-                        initialCustomModels={mergedCustomModels}
-                        customModelMetas={providerMetas[inst.provider] ?? {}}
-                        fetchedModels={inst.fetchedModels}
-                        fetchedAt={inst.fetchedAt}
-                        maskedKey={maskKey(inst.apiKey)}
-                        existingApiKey={inst.apiKey}
-                        onSave={(p) => handleSaveEdit(id, p)}
-                        onTest={(p) => handleTest(id, inst.provider, p)}
-                        testing={!!testingIds[id]}
-                        testStatus={result?.ok === true ? "success" : "idle"}
-                        onDelete={() => handleDelete(id)}
-                        onAddCustomModel={async (mid, meta) => {
-                          if (isCustom && cpId) {
-                            // Custom provider: the new model becomes part of the
-                            // provider's own model list (tools always true).
-                            await addCustomProviderModel(cpId, {
-                              id: mid,
-                              displayName: meta.displayName,
-                              vision: meta.vision,
-                              tools: true,
-                              maxContextTokens: meta.maxContextTokens,
-                            });
-                          } else {
-                            // Builtin: persist to BOTH the instance (back-compat) AND the provider pool.
-                            const nextInst = [...(inst.customModels ?? []), mid];
-                            await updateInstance(id, { customModels: nextInst });
-                            await addProviderCustomModel(inst.provider, mid);
-                            await setProviderCustomModelMeta(bp, mid, meta);
-                          }
-                          await reload();
-                        }}
-                        onUpdateCustomModelMeta={async (mid, meta) => {
-                          if (isCustom && cpId) {
-                            await updateCustomProviderModel(cpId, mid, {
-                              id: mid,
-                              displayName: meta.displayName,
-                              vision: meta.vision,
-                              tools: true,
-                              maxContextTokens: meta.maxContextTokens,
-                            });
-                          } else {
-                            await setProviderCustomModelMeta(bp, mid, meta);
-                          }
-                          await reload();
-                        }}
-                        onRemoveCustomModel={async (mid) => {
-                          if (isCustom && cpId) {
-                            await removeCustomProviderModel(cpId, mid);
-                          } else {
-                            // Remove from BOTH layers so the model truly disappears.
-                            const nextInst = (inst.customModels ?? []).filter((x) => x !== mid);
-                            await updateInstance(id, { customModels: nextInst });
-                            await removeProviderCustomModel(inst.provider, mid);
-                            await removeProviderCustomModelMeta(bp, mid); // cascade-clear pcmm
-                          }
-                          await reload();
-                        }}
-                        onRefreshModels={async (apiKey) => {
-                          // /v1/models is public — apiKey is optional (forwarded for parity).
-                          if (inst.provider !== "openrouter") return;
-                          const meta = getProviderMeta("openrouter")!;
-                          try {
-                            const fetched = await fetchOpenRouterModels(meta.defaultBaseUrl, apiKey || undefined);
-                            await updateInstance(id, { fetchedModels: fetched, fetchedAt: Date.now() });
-                            await reload();
-                          } catch {
-                            // silent for v1; user can retry
-                          }
-                        }}
-                      />
-                      {result?.ok === false && (
-                        <div
-                          className="mx-3.5 mb-3 rounded-chip border border-warning-line bg-warning-tint px-2.5 py-1.5 text-[11px] text-warning"
-                        >
-                          {t("customProvider.testFailed", { error: result.message })}
-                        </div>
-                      )}
-                    </>
-                  );
-                }}
-              />
-            </section>
-
-          </div>
-        ) : tab === "general" ? (
-          <div className="flex flex-col gap-7">
-            <section className="flex flex-col gap-3.5">
-              <div className="text-[15px] font-semibold tracking-[-0.005em] text-fg-1">{t("settings.language.sectionTitle")}</div>
-              <LanguageSelect />
-              <div className="flex flex-col gap-1.5">
-                <div className="font-mono text-[10px] uppercase tracking-[0.14em] text-fg-3">
-                  {t("settings.language.assistantLabel")}
-                </div>
-                <AssistantLanguageSelect />
-              </div>
-            </section>
-            <CdpInputSection
-              state={cdpInput}
-              onSet={async (next) => { setCdpInput(next); await setCdpInputEnabled(next); }}
-            />
-            <LocalBridgeSection />
-            <FeedbackSection instances={instances} />
-            <AboutSection />
-          </div>
-        ) : tab === "skills" ? (
-          <SkillsList onRunSkill={onRunSkill ?? (() => {})} />
-        ) : (
-          <SearchProviderSection />
-        )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function draftProviderMeta(
-  provider: ProviderRef,
-  options: ProviderTestOptions,
-): { id: ProviderRef; name: string; defaultBaseUrl: string; placeholder: string; models: ModelMeta[]; endpointVariants?: undefined } | null {
-  const baseUrl = options.baseUrl?.trim();
-  if (!provider.startsWith(CUSTOM_PREFIX) || !baseUrl || !/^https?:\/\//.test(baseUrl)) return null;
-  return {
-    id: provider,
-    name: options.providerName || provider,
-    defaultBaseUrl: baseUrl,
-    placeholder: "Custom",
-    models: options.candidateModels ?? [],
-    endpointVariants: undefined,
-  };
-}
-
-function SegmentedTabs({ value, onChange }: { value: Tab; onChange: (t: Tab) => void }) {
-  const t = useT();
-  const tabs: { id: Tab; label: string }[] = [
-    { id: "configs", label: t("settings.tabs.configs") },
-    { id: "skills", label: t("settings.tabs.skills") },
-    { id: "search", label: t("settings.tabs.search") },
-    { id: "general", label: t("settings.tabs.general") },
-  ];
-  return (
-    <div data-testid="settings-tabs" className="flex w-full overflow-hidden rounded-control border border-line">
-      {tabs.map((tb, i) => {
-        const active = value === tb.id;
-        return (
-          <button
-            key={tb.id}
-            onClick={() => onChange(tb.id)}
-            className={`flex-1 py-1.5 text-[12px] transition-colors ${i > 0 ? "border-l border-line" : ""} ${
-              active ? "bg-field font-medium text-fg-1" : "bg-transparent text-fg-2 hover:text-fg-1"
-            }`}
-          >
-            {tb.label}
-          </button>
-        );
-      })}
-    </div>
-  );
-}
-
-const MAX_LOG_CHARS = 100_000;
-// 留足 message≤4000 + env + JSON 转义开销，安全落在 256KB 请求体 bodyLimit 内。
-const MAX_LOG_BYTES = 150_000;
-
-export function FeedbackSection({ instances }: { instances: DecryptedInstance[] }) {
-  const t = useT();
-  const [message, setMessage] = useState("");
-  const [includeLogs, setIncludeLogs] = useState(false);
-  const [status, setStatus] = useState<"idle" | "sending" | "sent" | "error">("idle");
-
-  const active = instances[0];
-  const env: FeedbackEnv = {
-    version: chrome.runtime.getManifest().version,
-    userAgent: navigator.userAgent,
-    providerModel: active ? active.provider : "(no config)",
-    locale: getLocale(),
-  };
-  const managedApiKey = instances.find((i) => i.provider === "managed")?.apiKey;
-
-  async function onSend() {
-    const trimmed = message.trim();
-    if (!trimmed || status === "sending") return;
-    setStatus("sending");
-    try {
-      const logs = includeLogs
-        ? capLogBytes((await readRecentLogs(Date.now())).slice(0, MAX_LOG_CHARS), MAX_LOG_BYTES)
-        : undefined;
-      await submitFeedback({ message: trimmed, env, ...(logs ? { logs } : {}), ...(managedApiKey ? { apiKey: managedApiKey } : {}) });
-      setMessage("");
-      setIncludeLogs(false);
-      setStatus("sent");
-    } catch {
-      setStatus("error");
-    }
-  }
-
-  return (
-    <section className="flex flex-col gap-2.5">
-      <div className="text-[15px] font-semibold tracking-[-0.005em] text-fg-1">{t("settings.feedback.sectionTitle")}</div>
-      <p className="text-[12px] leading-[18px] text-fg-2">{t("settings.feedback.formHint")}</p>
-      <textarea
-        value={message}
-        onChange={(e) => { setMessage(e.target.value); if (status !== "idle") setStatus("idle"); }}
-        placeholder={t("settings.feedback.placeholder")}
-        rows={3}
-        className="w-full resize-y rounded-control border border-line bg-field px-2.5 py-2 text-[13px] text-fg-1 placeholder:text-fg-3 focus:outline-none"
-      />
-      <div className="flex items-center justify-between gap-3 pt-0.5">
-        <label className="flex items-center gap-2 text-[12px] text-fg-2">
-          <input type="checkbox" checked={includeLogs} onChange={(e) => setIncludeLogs(e.target.checked)} />
-          {t("settings.feedback.includeLogs")}
-        </label>
-        <div className="flex items-center gap-3">
-          {status === "sent" && <span className="text-[12px] text-success">{t("settings.feedback.sent")}</span>}
-          {status === "error" && <span className="text-[12px] text-warning">{t("settings.feedback.sendError")}</span>}
-          <button
-            onClick={onSend}
-            disabled={!message.trim() || status === "sending"}
-            className="shrink-0 rounded-control bg-accent px-3 py-1.5 text-[13px] font-medium text-canvas disabled:opacity-50"
-          >
-            {status === "sending" ? t("settings.feedback.sending") : t("settings.feedback.send")}
-          </button>
-        </div>
-      </div>
-      <div className="flex items-center gap-4 pt-0.5">
-        <a href={buildGithubNewIssueUrl(env)} target="_blank" rel="noopener noreferrer" className="text-[13px] font-medium text-accent hover:underline">{t("settings.feedback.githubButton")} ↗</a>
-        <a href={buildFeedbackMailto(env)} className="text-[13px] text-fg-2 hover:text-fg-1">{t("settings.feedback.emailButton")} ↗</a>
-      </div>
-    </section>
-  );
-}
-
-function CdpInputSection({
-  state,
-  onSet,
-}: {
-  state: boolean | undefined;
-  onSet: (next: boolean) => void;
-}) {
-  const t = useT();
-  const enabled = state === true;
-  return (
-    <section className="flex flex-col gap-3">
-      <div className="flex items-baseline justify-between">
-        <span className="text-[15px] font-semibold tracking-[-0.005em] text-fg-1">{t("settings.experimental")}</span>
-      </div>
-      <div className="flex flex-col gap-3 rounded-card border border-line bg-surface p-3.5">
-        <div className="flex items-start gap-3">
-          <div className="flex flex-1 flex-col gap-1">
-            <div className="text-[13px] font-medium text-fg-1">{t("settings.cdpInput.title")}</div>
-            <p className="text-[12px] leading-[18px] text-fg-2">
-              {t("settings.cdpInput.description")}
-            </p>
-            <p className="text-[11px] text-fg-3 mt-0.5">
-              {state === undefined
-                ? t("settings.cdpInput.statusNotAsked")
-                : enabled
-                ? t("settings.cdpInput.statusEnabled")
-                : t("settings.cdpInput.statusDisabled")}
-            </p>
-          </div>
-          <Switch checked={enabled} onChange={onSet} />
-        </div>
-        {enabled && (
-          <div className="flex flex-col gap-1.5 rounded-chip border border-warning-line bg-warning-tint px-3 py-2 text-[11px] leading-[16px] text-warning">
-            <span className="font-medium">{t("settings.cdpInput.warningTitle")}</span>
-            <ul className="flex flex-col gap-1 pl-3 text-warning/90">
-              <li className="list-['—__'] pl-0">{t("settings.cdpInput.warning1")}</li>
-              <li className="list-['—__'] pl-0">{t("settings.cdpInput.warning2")}</li>
-              <li className="list-['—__'] pl-0">{t("settings.cdpInput.warning3")}</li>
-            </ul>
-          </div>
-        )}
-      </div>
-    </section>
-  );
-}
-
-type BridgeStatus = { hasPermission: boolean; ready: boolean };
-
-// 向 SW 查活桥状态（nativeMessaging 是否授予 + 是否已连上 daemon）。SW 睡了/无响应
-// 时静默返回，不更新——面板保留上次已知状态。
-function queryBridgeStatus(cb: (s: BridgeStatus) => void): void {
-  try {
-    chrome.runtime.sendMessage({ type: "local-bridge:status" }, (res) => {
-      if (chrome.runtime.lastError) return;
-      if (res && typeof res.hasPermission === "boolean") cb(res as BridgeStatus);
-    });
-  } catch {
-    /* noop */
-  }
-}
-
-type PanelAgent = { id: string; label: string; installed: boolean; enabled: boolean };
-
-// Settings「本地 Agent」列表 — 一次性查询，无轮询（挂载/桥就绪/开关交互后各触发一次）。
-function queryLocalAgents(cb: (agents: PanelAgent[]) => void): void {
-  try {
-    chrome.runtime.sendMessage({ type: "local-agents:list" }, (res) => {
-      if (chrome.runtime.lastError) return;
-      if (res && Array.isArray(res.agents)) cb(res.agents as PanelAgent[]);
-    });
-  } catch {
-    /* noop */
-  }
-}
-
-// Settings「本地打通」— 已授权 skill grants 一次性查询（无轮询）。
-function queryGrants(cb: (grants: GrantRecord[]) => void): void {
-  try {
-    chrome.runtime.sendMessage({ type: "local-grants:list" }, (res) => {
-      if (chrome.runtime.lastError) return;
-      if (res && Array.isArray(res.grants)) cb(res.grants as GrantRecord[]);
-    });
-  } catch {
-    /* noop */
-  }
-}
-
-// Settings「本地打通」— 最近脚本执行审计一次性查询（无轮询）。
-function queryAudit(cb: (entries: AuditEntry[]) => void): void {
-  try {
-    chrome.runtime.sendMessage({ type: "local-audit:list" }, (res) => {
-      if (chrome.runtime.lastError) return;
-      if (res && Array.isArray(res.entries)) cb(res.entries as AuditEntry[]);
-    });
-  } catch {
-    /* noop */
-  }
-}
-
-// 本地打通开关 + 实时状态。开=请求 nativeMessaging（用户手势）→ SW onAdded 连桥；
-// 关=移除权限 → SW onRemoved 断桥。挂载期每 1.5s 轮询一次状态（连接是异步的）。
-export function LocalBridgeSection() {
-  const t = useT();
-  const [status, setStatus] = useState<BridgeStatus | null>(null);
-  const [agents, setAgents] = useState<PanelAgent[]>([]);
-  const [failedId, setFailedId] = useState<string | null>(null);
-  const [grants, setGrants] = useState<GrantRecord[]>([]);
-  const [audit, setAudit] = useState<AuditEntry[]>([]);
-  const [auditOpen, setAuditOpen] = useState(false);
-
-  useEffect(() => {
-    queryBridgeStatus(setStatus);
-    const id = setInterval(() => queryBridgeStatus(setStatus), 1500);
-    return () => clearInterval(id);
-  }, []);
-
-  useEffect(() => {
-    if (status?.ready) queryLocalAgents(setAgents);
-    else setAgents([]);
-  }, [status?.ready]);
-
-  useEffect(() => {
-    if (status?.ready) {
-      queryGrants(setGrants);
-      queryAudit(setAudit);
-    } else {
-      setGrants([]);
-      setAudit([]);
-    }
-  }, [status?.ready]);
-
-  const onRevoke = (key: string) => {
-    try {
-      chrome.runtime.sendMessage({ type: "local-grants:revoke", key }, (res) => {
-        if (chrome.runtime.lastError) return;
-        if (res?.ok) queryGrants(setGrants);
-      });
-    } catch {
-      /* noop */
-    }
-  };
-
-  const enabled = status?.hasPermission ?? false;
-  const onToggle = async (next: boolean) => {
-    try {
-      if (next) await chrome.permissions.request({ permissions: ["nativeMessaging"] });
-      else await chrome.permissions.remove({ permissions: ["nativeMessaging"] });
-    } catch {
-      /* 用户取消了权限弹窗 */
-    }
-    queryBridgeStatus(setStatus);
-  };
-
-  const onAgentToggle = (id: string, next: boolean) => {
-    setFailedId(null);
-    try {
-      chrome.runtime.sendMessage({ type: "local-agents:toggle", id, next }, (res) => {
-        if (chrome.runtime.lastError) return;
-        if (res?.ok) queryLocalAgents(setAgents);
-        else setFailedId(id);
-      });
-    } catch {
-      /* noop */
-    }
-  };
-
-  const statusText =
-    status == null
-      ? ""
-      : !status.hasPermission
-        ? t("settings.localBridge.statusOff")
-        : status.ready
-          ? t("settings.localBridge.statusConnected")
-          : t("settings.localBridge.statusEnabledNotConnected");
-
-  return (
-    <section className="flex flex-col gap-3">
-      <div className="flex items-baseline justify-between">
-        <span className="text-[15px] font-semibold tracking-[-0.005em] text-fg-1">
-          {t("settings.localBridge.sectionTitle")}
-        </span>
-      </div>
-      <div className="flex flex-col gap-3 rounded-card border border-line bg-surface p-3.5">
-        <div className="flex items-start justify-between gap-3">
-          <div className="flex flex-col gap-1">
-            <div className="text-[13px] font-medium text-fg-1">{t("settings.localBridge.title")}</div>
-            <div className="text-[12px] leading-relaxed text-fg-3">{t("settings.localBridge.description")}</div>
-            {statusText && (
-              <div className={`text-[12px] ${status?.ready ? "text-fg-1" : "text-fg-3"}`}>{statusText}</div>
-            )}
-          </div>
-          <Switch checked={enabled} onChange={onToggle} />
-        </div>
-        {status?.ready && agents.length > 0 && (
-          <div className="flex flex-col gap-2 border-t border-line pt-3">
-            <div className="text-[12px] font-medium text-fg-2">{t("settings.localBridge.agentsTitle")}</div>
-            {agents.map((a) => (
-              <div key={a.id} className="flex flex-col gap-1">
-                <div className="flex items-center justify-between gap-3">
-                  <div className="flex items-baseline gap-2">
-                    <span className="text-[13px] text-fg-1">{a.label}</span>
-                    {!a.installed && (
-                      <span className="text-[11px] text-fg-3">{t("settings.localBridge.agentNotInstalled")}</span>
-                    )}
-                  </div>
-                  <Switch checked={a.enabled} onChange={(next) => onAgentToggle(a.id, next)} />
-                </div>
-                {failedId === a.id && (
-                  <div className="text-[11px] text-fg-3">{t("settings.localBridge.agentEnableFailed")}</div>
-                )}
-              </div>
-            ))}
-          </div>
-        )}
-        {status?.ready && grants.length > 0 && (
-          <div className="flex flex-col gap-2 border-t border-line pt-3">
-            <div className="text-[12px] font-medium text-fg-2">{t("settings.localBridge.grantsTitle")}</div>
-            {grants.map((g) => (
-              <div key={g.key} className="flex items-start justify-between gap-3">
-                <div className="flex flex-col gap-0.5">
-                  <span className="text-[13px] text-fg-1">{g.skillName}</span>
-                  <span className="text-[11px] text-fg-3">
-                    {g.envelope.runnableScripts.join(", ")}
-                    {g.envelope.allowedDomains.length > 0 && ` · ${g.envelope.allowedDomains.join(", ")}`}
-                    {g.envelope.extraWrites.length > 0 && ` · ${g.envelope.extraWrites.join(", ")}`}
-                  </span>
-                  <span className="text-[11px] text-fg-3">
-                    {new Date(g.grantedAt).toLocaleDateString()}
-                  </span>
-                </div>
-                <button
-                  type="button"
-                  onClick={() => onRevoke(g.key)}
-                  className="shrink-0 rounded border border-line px-2 py-0.5 text-[11px] text-fg-2 hover:text-fg-1"
-                >
-                  {t("settings.localBridge.revoke")}
-                </button>
-              </div>
-            ))}
-          </div>
-        )}
-        {status?.ready && audit.length > 0 && (
-          <div className="flex flex-col gap-2 border-t border-line pt-3">
-            <button
-              type="button"
-              onClick={() => setAuditOpen((v) => !v)}
-              className="self-start text-[12px] font-medium text-fg-2 hover:text-fg-1"
-            >
-              {t("settings.localBridge.auditTitle")} {auditOpen ? "▾" : "▸"}
-            </button>
-            {auditOpen &&
-              audit.map((e, i) => (
-                <div key={`${e.ts}-${i}`} className="flex items-baseline justify-between gap-2 text-[11px]">
-                  <span className="truncate text-fg-1">
-                    {e.skillName} · {e.entry}
-                  </span>
-                  <span className="shrink-0 text-fg-3">
-                    {e.exitCode === 0 && !e.timedOut
-                      ? t("settings.localBridge.auditOk")
-                      : t("settings.localBridge.auditFailed")}
-                    {" · "}
-                    {new Date(e.ts).toLocaleString()}
-                  </span>
-                </div>
-              ))}
-          </div>
-        )}
-      </div>
-    </section>
-  );
-}
-
-function Switch({
-  checked,
-  onChange,
-}: {
-  checked: boolean;
-  onChange: (next: boolean) => void;
-}) {
-  return (
-    <button
-      type="button"
-      role="switch"
-      aria-checked={checked}
-      onClick={() => onChange(!checked)}
-      className={`relative inline-flex h-6 w-11 flex-shrink-0 items-center rounded-full border transition-colors ${
-        checked ? "border-accent-line bg-accent-tint" : "border-line bg-field"
-      }`}
-    >
-      <span
-        className={`inline-block h-3.5 w-3.5 transform rounded-full transition-transform ${
-          checked ? "translate-x-6 bg-accent" : "translate-x-1 bg-fg-3"
-        }`}
-      />
-    </button>
-  );
-}
-
-function AboutSection() {
-  const t = useT();
-  const v = chrome.runtime.getManifest().version;
-  return (
-    <section className="flex flex-col gap-3.5">
-      <div className="h-px w-full bg-line" />
-      <div className="flex items-center gap-2.5">
-        <img
-          src={chrome.runtime.getURL("icons/icon-128.png")}
-          alt="Pie"
-          className="h-[26px] w-[26px] flex-shrink-0 rounded-chip"
-        />
-        <div className="flex flex-col gap-0.5">
-          <div className="flex items-baseline gap-1.5">
-            <span className="text-[13px] font-semibold text-fg-1">Pie</span>
-            <span className="font-mono text-[11px] text-fg-2">v{v}</span>
-          </div>
-          <span className="text-[11px] text-fg-3">{t("settings.about.tagline")}</span>
-        </div>
-        <div className="flex-1" />
-        <div className="flex items-center gap-3">
-          <a href="https://www.pie.chat/" target="_blank" rel="noopener noreferrer" className="text-[12px] text-fg-2 hover:text-fg-1">{t("settings.about.website")} ↗</a>
-          <a href="https://github.com/WiseriaAI/pie-ai-agent/releases" target="_blank" rel="noopener noreferrer" className="text-[12px] text-fg-2 hover:text-fg-1">{t("settings.about.changelog")} ↗</a>
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function maskKey(k: string): string {
-  return k.length <= 8 ? "••••••••" : `${k.slice(0, 4)}...${k.slice(-4)}`;
-}
